feat(diri): add `order_type` field
