+++ crates/diri/src/storage/types.rs
@@ -5,7 +5,7 @@ use starknet::core::utils::parse_cairo_short_string;
 use std::fmt::LowerHex;
 
 use crate::orderbook::events::{
-    OrderCancelled, OrderExecuted, OrderFulfilled, OrderPlaced, RollbackStatus, U256,
+    OrderCancelled, OrderExecuted, OrderFulfilled, RollbackStatus, U256,
 };
 
 #[derive(Debug, Clone)]
@@ -31,36 +31,10 @@ pub struct PlacedData {
     pub broker_id: String,
 }
 
-impl From<OrderPlaced> for PlacedData {
-    fn from(value: OrderPlaced) -> Self {
-        match value {
-            OrderPlaced::V1(value) => Self {
-                order_hash: to_hex_str(&value.order_hash),
-                order_version: to_hex_str(&value.order_version),
-                order_type: format!("{:?}", value.order_type),
-                cancelled_order_hash: to_hex_str_opt(&value.cancelled_order_hash),
-                route: format!("{:?}", value.order.route),
-                currency_address: to_hex_str(&Felt::from(value.order.currency_address)),
-                currency_chain_id: to_hex_str(&value.order.currency_chain_id),
-                salt: to_hex_str(&value.order.salt),
-                offerer: to_hex_str(&Felt::from(value.order.offerer)),
-                token_chain_id: format!("0x{:x}", value.order.token_chain_id),
-                token_address: to_hex_str(&Felt::from(value.order.token_address)),
-                token_id: u256_to_hex_opt(&value.order.token_id),
-                quantity: u256_to_hex(&value.order.quantity),
-                start_amount: u256_to_hex(&value.order.start_amount),
-                end_amount: u256_to_hex(&value.order.end_amount),
-                start_date: value.order.start_date,
-                end_date: value.order.end_date,
-                broker_id: to_hex_str(&Felt::from(value.order.broker_id)),
-            },
-        }
-    }
-}
-
 #[derive(Debug, Clone)]
 pub struct CancelledData {
     pub order_hash: String,
+    pub order_type: String,
     pub reason: String,
 }
 
@@ -69,6 +43,7 @@ impl From<OrderCancelled> for CancelledData {
         match value {
             OrderCancelled::V1(value) => Self {
                 order_hash: to_hex_str(&value.order_hash),
+                order_type: format!("{:?}", value.order_type),
                 reason: parse_cairo_short_string(&value.reason)
                     .unwrap_or(to_hex_str(&value.reason)),
             },
@@ -79,6 +54,7 @@ impl From<OrderCancelled> for CancelledData {
 #[derive(Debug, Clone)]
 pub struct RollbackStatusData {
     pub order_hash: String,
+    pub order_type: String,
     pub reason: String,
 }
 
@@ -87,6 +63,7 @@ impl From<RollbackStatus> for RollbackStatusData {
         match value {
             RollbackStatus::V1(value) => Self {
                 order_hash: to_hex_str(&value.order_hash),
+                order_type: format!("{:?}", value.order_type),
                 reason: parse_cairo_short_string(&value.reason)
                     .unwrap_or(to_hex_str(&value.reason)),
             },
@@ -97,6 +74,7 @@ impl From<RollbackStatus> for RollbackStatusData {
 #[derive(Debug, Clone)]
 pub struct FulfilledData {
     pub order_hash: String,
+    pub order_type: String,
     pub fulfiller: String,
     pub related_order_hash: Option<String>,
 }
@@ -109,6 +87,7 @@ impl From<OrderFulfilled> for FulfilledData {
 
                 Self {
                     order_hash: to_hex_str(&value.order_hash),
+                    order_type: format!("{:?}", value.order_type),
                     fulfiller: to_hex_str(&Felt::from(value.fulfiller)),
                     related_order_hash: to_hex_str_opt(&related_order_hash),
                 }
@@ -121,6 +100,7 @@ impl From<OrderFulfilled> for FulfilledData {
 pub struct ExecutedData {
     pub version: u8,
     pub order_hash: String,
+    pub order_type: Option<String>,
     pub transaction_hash: Option<String>,
     pub from: Option<String>,
     pub to: Option<String>,
@@ -132,6 +112,7 @@ impl From<OrderExecuted> for ExecutedData {
             OrderExecuted::V0(v) => Self {
                 version: 0,
                 order_hash: to_hex_str(&v.order_hash),
+                order_type: None,
                 transaction_hash: None,
                 from: None,
                 to: None,
@@ -139,13 +120,15 @@ impl From<OrderExecuted> for ExecutedData {
             OrderExecuted::V1(v) => Self {
                 version: 1,
                 order_hash: to_hex_str(&v.order_hash),
+                order_type: None,
                 transaction_hash: Some(to_hex_str(&v.transaction_hash)),
                 from: Some(to_hex_str(&Felt::from(v.from))),
                 to: Some(to_hex_str(&Felt::from(v.to))),
             },
             OrderExecuted::V2(v) => Self {
-                version: 1,
+                version: 2,
                 order_hash: to_hex_str(&v.order_hash),
+                order_type: Some(format!("{:?}", v.order_type)),
                 transaction_hash: Some(to_hex_str(&v.transaction_hash)),
                 from: Some(to_hex_str(&Felt::from(v.from))),
                 to: Some(to_hex_str(&Felt::from(v.to))),
